Make the UI poll for data and POST actions
There is a race condition in here but it rarely shows up given that this is only
controlled on a LAN where the latency is very low and errors are corrected
within the poll window of 1 second.

On rare occasions, the animation will give a slight wobble effect. This is fine,
however one way of fixing this is with a flag that disables any updates from the
polling code for the time between the button being clicked and a response coming
back from the POST request.