Separate LED and Relay and add init indicator
I'm using the "state" as a logical flag, such that:
 true  = relay on
 false = relay off

However, the relay board activates when the input is pulled LOW, such that:
 LOW   = relay on
 HIGH  = relay off

This means the same pin can't drive the LED and the relay, as their states need
to be opposites.

I've also made the LED flash while it's connecting to the wifi network.